+++ src/app/(private)/layout.tsx
@@ -1,5 +1,5 @@
 import Navbar from '@/components/dashboard/navbar'
-import Sidenav from '@/components/dashboard/sidenav'
+import Sidebar from '@/components/dashboard/sidebar'
 
 export default function RootLayout({
   children,
@@ -8,7 +8,7 @@ export default function RootLayout({
     <div className="overflow-hidden h-screen">
       <Navbar />
       <div className="grid w-full md:grid-cols-[220px_1fr] lg:grid-cols-[280px_1fr] h-[calc(100dvh-3.5rem)]">
-        <Sidenav />
+        <Sidebar />
         <main className="flex flex-1 flex-col gap-4 p-4 lg:gap-6 lg:p-6 overflow-auto">
           {children}
         </main>
