+++ src/components/dashboard/navbar.tsx
@@ -5,8 +5,8 @@ import { Button } from '../ui/button'
 import Menu from './menu'
 import { MenuIcon } from 'lucide-react'
 import { Sheet, SheetContent, SheetTrigger } from '../ui/sheet'
-import { ModeToggle } from './mode-toggle'
-import LanguageToggle from './language-toggle'
+import { ModeToggle } from './modeToggle'
+import LanguageToggle from './languageToggle'
 import { Avatar, AvatarFallback } from '../ui/avatar'
 import { LuUser2 } from 'react-icons/lu'
 import {
